Load contrast database

Load a pre-calculated, organism-specific contrast database and return it as a SummarizedExperiment.

loadContrastDatabase(
  organism = c("Human", "Mouse"),
  mode = c("ANALYSIS", "DEMO"),
  mustWork = TRUE
)

Arguments

organism

Character scalar selecting the organism for which to load the contrast database. One of "Human" or "Mouse".

mode

When in "ANALYSIS" mode (default) the complete contrast DB is queried. "DEMO" mode employs a small "toy" database for the queries. "DEMO" should only be used for testing/demonstration purposes and never for actual analysis purposes.

mustWork

Logical scalar. If FALSE and the contrast database is not available, return an empty SummarizedExperiment object. If TRUE (the default) and the contrast database is not available, loadContrastDatabase throws an error.

Value

A SummarizedExperiment with pre-calculated contrasts as assays.

Details

Organism-specific databases are compiled in HDF5SummarizedExperiment objects. The first time `loadContrastDatabase()` is called for a database either directly or via `queryWithContrasts()` the required objects will be automatically downloaded from `ExperimentHub` and cached in the user ExperimentHub directory (see `ExperimentHub::getExperimentHubOption("CACHE")`) using the `orthosData` companion data-package.

Author

Panagiotis Papasaikas, Michael Stadler

Examples

# \donttest{

# !!!Note!!! mode="DEMO" for demonstration purposes only. Default is mode="ANALYSIS"
SE_mouse_demoDB <- loadContrastDatabase (organism="Mouse", mode="DEMO")
#> Error in loadContrastDatabase(organism = "Mouse", mode = "DEMO"): could not find function "loadContrastDatabase"
SE_mouse_demoDB
#> Error in eval(expr, envir, enclos): object 'SE_mouse_demoDB' not found

SE_human_demoDB <- loadContrastDatabase (organism="Human", mode="DEMO")
#> Error in loadContrastDatabase(organism = "Human", mode = "DEMO"): could not find function "loadContrastDatabase"
SE_human_demoDB
#> Error in eval(expr, envir, enclos): object 'SE_human_demoDB' not found

# }