Load contrast database

Load a pre-calculated, organism-specific contrast database and return it as a SummarizedExperiment.

loadContrastDatabase(
  organism = c("Human", "Mouse"),
  mode = c("ANALYSIS", "DEMO"),
  mustWork = TRUE
)

Arguments

organism

Character scalar selecting the organism for which to load the contrast database. One of "Human" or "Mouse".

mode

When in "ANALYSIS" mode (default) the complete contrast DB is queried. "DEMO" mode employs a small "toy" database for the queries. "DEMO" should only be used for testing/demonstration purposes and never for actual analysis purposes.

mustWork

Logical scalar. If FALSE and the contrast database is not available, return an empty SummarizedExperiment object. If TRUE (the default) and the contrast database is not available, loadContrastDatabase throws an error.

Value

A SummarizedExperiment with pre-calculated contrasts as assays.

Details

Organism-specific databases are compiled in HDF5SummarizedExperiment objects. The first time `loadContrastDatabase()` is called for a database either directly or via `queryWithContrasts()` the required objects will be automatically downloaded from `ExperimentHub` and cached in the user ExperimentHub directory (see `ExperimentHub::getExperimentHubOption("CACHE")`) using the `orthosData` companion data-package.

Author

Panagiotis Papasaikas, Michael Stadler

Examples

# \donttest{

# !!!Note!!! mode="DEMO" for demonstration purposes only. Default is mode="ANALYSIS"
SE_mouse_demoDB <- loadContrastDatabase (organism="Mouse", mode="DEMO")
#> demo_decomposed_contrasts_mouse_rds  already present in cache at: /Users/runner/Library/Caches/org.R-project.R/R/ExperimentHub/mouse_v212_NDF_c100_DEMOse.rds
#> demo_decomposed_contrasts_mouse_hdf5  already present in cache at: /Users/runner/Library/Caches/org.R-project.R/R/ExperimentHub/mouse_v212_NDF_c100_DEMOassays.h5
SE_mouse_demoDB
#> class: SummarizedExperiment 
#> dim: 20339 988 
#> metadata(0):
#> assays(4): INPUT_CONTRASTS DECODED_CONTRASTS RESIDUAL_CONTRASTS CONTEXT
#> rownames(20339): 0610005C13Rik 0610009B22Rik ... n-R5s146 n-R5s149
#> rowData names(15): seqnames start ... ENTREZ_GENE_ID ARCHS4_ID
#> colnames(988): GSM1053282 GSM1061179 ... GSM970471 GSM984548
#> colData names(33): aligned_reads channel_count ... HasAssignedCNT
#>   CNTname

SE_human_demoDB <- loadContrastDatabase (organism="Human", mode="DEMO")
#> demo_decomposed_contrasts_human_rds  already present in cache at: /Users/runner/Library/Caches/org.R-project.R/R/ExperimentHub/human_v212_NDF_c100_DEMOse.rds
#> demo_decomposed_contrasts_human_hdf5  already present in cache at: /Users/runner/Library/Caches/org.R-project.R/R/ExperimentHub/human_v212_NDF_c100_DEMOassays.h5
SE_human_demoDB
#> class: SummarizedExperiment 
#> dim: 20411 1000 
#> metadata(0):
#> assays(4): INPUT_CONTRASTS DECODED_CONTRASTS RESIDUAL_CONTRASTS CONTEXT
#> rownames(20411): A1BG A1BG-AS1 ... WFDC5 XAGE2
#> rowData names(15): seqnames start ... ENTREZ_GENE_ID ARCHS4_ID
#> colnames(1000): GSM1115210 GSM1202518 ... GSM6505602 GSM6573314
#> colData names(33): aligned_reads channel_count ... HasAssignedCNT
#>   CNTname

# }